fix modules copying in certain cases
since the destination is specified instead of the actual file, cp
probably does not check if the file exists but rather the destination
(directory) and does not copy it.

what's strange is that it does happen if used from GameOn Live
CD/DVD/USB environment, notably if done via chroot but it does happen if
used after GameOn is installed.

Signed-off-by: Ivailo Monev <xakepa10@laimg.moc>